+++ funcs_aux/__init__.py
@@ -1,4 +1,4 @@
 from .arrays import array_2d_get_compact_str
 from .iterables import Iterables
-from .results import ResultSucceedSimple, ResultFull
+from .results import ResultValue, ResultFunc
 from .strings import Strings