+++ test__strings.py
@@ -1,7 +1,7 @@
 import pytest
 
 from funcs_aux import *
-from funcs_aux import ResultSucceedSimple
+from funcs_aux import ResultValue
 
 
 # =====================================================================================================================